Show and Index actions for page work now
sh files for both index and show were created to be used for testing
To 'show' an individual page I used params to check for the page being
requested.
Index has basic functionality for page. You can view all pages from
every user, but not all pages that the user created. This is a UX
problem that must be fixed in the future.