Improved functionality of create in post
Create was only creating a post on the first blog regardless of which
blog specified in params.  It now works properly.
Authorization is still a problem for anyone can create a post on anyone
else's blog. Users should only be able to create a post on a blog that
belongs to them.